+++ src/ServerTelemetryChannel/Implementation/SamplingPercentageEstimatorTelemetryProcessor.cs
@@ -5,7 +5,6 @@
 
     using Microsoft.ApplicationInsights.Channel;
     using Microsoft.ApplicationInsights.Extensibility;
-    using Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel.Implementation;
 
     /// <summary>
     /// Represents a method that is invoked every time sampling percentage is evaluated